refactor: organize insight cards into card folders
