+++ frontend/src/insights/InsightsPage.tsx
@@ -1,17 +1,17 @@
 import { useState } from 'react'
 import { useAuth } from '@/hooks/useAuth'
-import { CashFlowCard } from './components/CashFlowCard'
+import { CashFlowCard } from './components/cash-flow-card/CashFlowCard'
 import {
   IncomeExpenseBreakdownCard,
   type BreakdownMode,
-} from './components/IncomeExpenseBreakdownCard'
+} from './components/income-expense-breakdown-card/IncomeExpenseBreakdownCard'
 import { InsightsFloatingRangeControl } from './components/InsightsFloatingRangeControl'
-import { MerchantDistributionCard } from './components/MerchantDistributionCard'
-import { MerchantRankingCard } from './components/MerchantRankingCard'
-import { FundFlowCard } from './components/FundFlowCard'
-import { NetWorthCard } from './components/NetWorthCard'
-import { PeriodGlanceCard } from './components/PeriodGlanceCard'
-import { SavingsRateTrendCard } from './components/SavingsRateTrendCard'
+import { MerchantDistributionCard } from './components/merchant-distribution-card/MerchantDistributionCard'
+import { MerchantRankingCard } from './components/merchant-ranking-card/MerchantRankingCard'
+import { FundFlowCard } from './components/fund-flow-card/FundFlowCard'
+import { NetWorthCard } from './components/net-worth-card/NetWorthCard'
+import { PeriodGlanceCard } from './components/period-glance-card/PeriodGlanceCard'
+import { SavingsRateTrendCard } from './components/savings-rate-trend-card/SavingsRateTrendCard'
 import { useInsightsCardData } from './hooks/useInsightsCardData'
 import { useInsightsCardQueries } from './hooks/useInsightsCardQueries'
 import { useInsightsCardVisibilityMap } from './hooks/useInsightsCardVisibilityMap'
