+++ frontend/src/insights/components/cash-flow-card/CashFlowBarChart.tsx
