+++ frontend/src/insights/components/fund-flow-card/FundFlowCategoryList.tsx
