+++ frontend/src/insights/components/fund-flow-card/FundFlowChart.tsx
