+++ frontend/src/insights/components/income-expense-breakdown-card/IncomeExpenseTrendSections.tsx
