+++ frontend/src/insights/components/merchant-distribution-card/MerchantDistributionLegend.tsx
