+++ frontend/src/insights/components/net-worth-card/NetWorthChart.tsx
