+++ frontend/src/insights/components/period-glance-card/PeriodGlancePrimaryPanel.tsx
