+++ frontend/src/insights/components/period-glance-card/PeriodGlanceSupportGrid.tsx
