+++ frontend/src/insights/components/period-glance-card/periodGlanceDisplay.ts
