+++ frontend/src/insights/components/period-glance-card/useFittedPrimaryAmount.ts
