+++ frontend/src/insights/components/savings-rate-trend-card/SavingsRateChart.tsx
